Added /audio

+++ distortBot.py.py
@@ -6,6 +6,7 @@ import sys
 from multiprocessing import Pool
 import subprocess
 import psutil
+import ffmpeg
 
 TOKEN = ""
 GET_GIF = 0
@@ -13,12 +14,12 @@ GET_GIF = 0
 # Start command that explains the usage of the bot
 def start(update, context):     
 
-    update.message.reply_text("This is a distortion bot. Use /gif and wait for a reply.")
+    update.message.reply_text("This is a distortion bot. Use /help to see all the commands")
 
 # Help command that explains the usage of the bot
 def help(update, context):
 
-    update.message.reply_text("Use /gif and wait for a reply. I'll create a distorted GIF of it")
+    update.message.reply_text("/gif - Send me a picture and I'll create a distorted gif on it\n/audio - Reply an audio and I'll distort it")
 
 # Distort function using imagemmagick
 def distort(imageIn, imageOut, dims, pct, directory, user_id):
@@ -27,6 +28,10 @@ def distort(imageIn, imageOut, dims, pct, directory, user_id):
     if imageOut in os.listdir(directory): return
     print(f"{pct}%: {imageIn} -> {imageOut}")
     os.system(f"magick toDistort{user_id}.jpg -liquid-rescale {pct}x{pct}%! -resize {dims[1]}x{dims[0]}! {imageOut}")
+
+def vibrato(audioIn, audioOut, pct=12):
+    print(f"{pct}%: {audioIn} -> {audioOut}")
+    os.system(f"ffmpeg -i {audioIn} -filter_complex \"vibrato=f={pct}\" {audioOut}")
 
 # Gif command to reply and return to GET_GIF
 def gif(update, context):
@@ -43,20 +48,36 @@ def enableMultithread(function, args):
     pool.close()
     pool.join()
 
-def downloadImages(update, context, pic, user_id):
+def downloadImage(update, context, pic, user_id):
     
      # First message to the bot
     if pic:
         imageIn = update.message.photo[-1].file_id
         image = context.bot.get_file(file_id = imageIn)
         image.download(f"toDistort{user_id}.jpg")
-        update.message.reply_text("We are processing your image. Wait for it!")
+        update.message.reply_text("yOuR pHoTO iS bEinG dIsTorRTeD")
         return image
     # If is not the user's first message to the bot
     else: 
         update.message.reply_text("Send me a pic :)")
         if update.message.photo:
-            update.message.reply_text("We are processing your image. Wait for it!")
+            update.message.reply_text("yOuR pHoTO iS bEinG dIsTorRTeD")
+
+def downloadAudio(update, context, audio, user_id):
+
+    if not update.message.reply_to_message.voice:
+        update.message.reply_text("tHiS Is nOT aN AUdIo")
+        return
+
+    audio  = context.bot.getFile(update.message.reply_to_message.voice)
+    ogg = f"audio{user_id}.ogg"
+    wav = f"audio{user_id}.wav"
+    audio.download(ogg)
+
+    stream = ffmpeg.input(ogg)
+    stream = ffmpeg.output(stream, wav)
+    stream = ffmpeg.overwrite_output(stream)
+    return stream
 
 def createMP4(update, context, directory, user_id):
     
@@ -73,12 +94,15 @@ def createMP4(update, context, directory, user_id):
     video = open(f"distorted{user_id}.mp4", "rb")
     return video
 
-def deleteDirs(directory, file1, file2, user_id):
+def deleteDirs(directory, file1, file2, file3, user_id):
 
     # Deleting the created and downloaded files
-    os.system(f"rm -rf {directory}")
+    if directory:
+        os.system(f"rm -rf {directory}")
     os.remove(f"{file1[:-4]}{user_id}{file1[len(file1) -  4::]}")
     os.remove(f"{file2[:-4]}{user_id}{file2[len(file2) -  4::]}")
+    if file3:
+        os.remove(f"{file3[:-4]}{user_id}{file3[len(file3) -  4::]}")
 
 # Creating the gif
 def getGif(update, context):
@@ -86,7 +110,7 @@ def getGif(update, context):
     user_id = update.message.from_user.id
     pic = update.message.photo
 
-    downloadImages(update, context, pic, user_id)
+    downloadImage(update, context, pic, user_id)
 
     # Getting the image dimmensions and setting up a list for the 60 arguments of distort function
     dims = imageio.imread(open(f"toDistort{user_id}.jpg", "rb")).shape
@@ -110,7 +134,24 @@ def getGif(update, context):
     context.bot.sendAnimation(animation = video, chat_id = update.message.chat_id)
     video.close()
 
-    deleteDirs(directory, "toDistort.jpg", "distorted.mp4", user_id)
+    deleteDirs(directory, "toDistort.jpg", "distorted.mp4", None, user_id)
+
+
+def audio(update, context):
+
+    user_id = update.message.from_user.id
+    audio = update.message.audio
+
+    audioIn = downloadAudio(update, context, audio, user_id)
+    ffmpeg.run(audioIn)
+    audioOut = f"audioDistorted{user_id}.wav"
+    vibrato(f"audio{user_id}.wav", audioOut)
+    voiceToSend = open(audioOut, "rb")
+    context.bot.sendVoice(voice = voiceToSend, chat_id = update.message.chat_id)
+    voiceToSend.close()
+
+    deleteDirs(None, "audio.wav", "audioDistorted.wav", "audio.ogg", user_id)
+
 
 def main():
     logger = logging.getLogger(__name__)
@@ -122,13 +163,15 @@ def main():
 
     dp.add_handler(CommandHandler("start", start))
     dp.add_handler(CommandHandler("help", help))
+    dp.add_handler(CommandHandler("audio", audio))
     dp.add_handler(ConversationHandler(
         entry_points = [CommandHandler("gif", gif)],
         states = {
-            GET_GIF: [MessageHandler(Filters.all, getGif)]
+            GET_GIF: [MessageHandler(Filters.photo | Filters.regex(r'\/gif') | Filters.regex(r'\/gif@theDistortionBot'), getGif)]
         },
         fallbacks = []
     ))
+
 
     updater.start_polling()
     logging.info("=== Bot running! ===")
